updated-backend/park_easy/urls.py
from django.urls import path, include
from rest_framework import routers

from park_easy.views import MyObtainTokenPairView, RegisterView, UserViewSet
from rest_framework_simplejwt.views import TokenRefreshView

router = routers.DefaultRouter()
router.register(r'user', UserViewSet)
urlpatterns = [
    path('api-auth/', include('rest_framework.urls')),
    path('', include(router.urls)),
    path('login/', MyObtainTokenPairView.as_view(), name='token_obtain_pair'),
    path('login/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('register/', RegisterView.as_view(), name='auth_register'),
    path('', include('parkinglot.urls')),
]
